Clean up IncludeDirectoryResolver and CastXml
