+++ SharpGen.UnitTests/IncludeDirectoryResolverTests.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Runtime.InteropServices;
-using System.Text;
 using SharpGen.Config;
 using SharpGen.Logging;
 using SharpGen.Parser;
@@ -23,7 +20,7 @@ namespace SharpGen.UnitTests
             Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "Registry paths can only be resolved on Windows");
 
             var resolver = new IncludeDirectoryResolver(Logger);
-            resolver.AddDirectories(new[] { new IncludeDirRule(@"=HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows Kits\Installed Roots\KitsRoot10") });
+            resolver.AddDirectories(new IncludeDirRule(@"=HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows Kits\Installed Roots\KitsRoot10"));
 
             Assert.Equal($@"-isystem""C:\Program Files (x86)\Windows Kits\10""", resolver.IncludePaths.First());
         }
@@ -34,7 +31,7 @@ namespace SharpGen.UnitTests
             Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "Registry paths can only be resolved on Windows");
 
             var resolver = new IncludeDirectoryResolver(Logger);
-            resolver.AddDirectories(new[] { new IncludeDirRule(@"=HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows Kits\Installed Roots\KitsRoot10;SubFolder") });
+            resolver.AddDirectories(new IncludeDirRule(@"=HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows Kits\Installed Roots\KitsRoot10;SubFolder"));
 
             Assert.Equal($@"-isystem""C:\Program Files (x86)\Windows Kits\10\SubFolder""", resolver.IncludePaths.First());
         }
@@ -45,7 +42,7 @@ namespace SharpGen.UnitTests
             Skip.If(RuntimeInformation.IsOSPlatform(OSPlatform.Windows));
 
             var resolver = new IncludeDirectoryResolver(Logger);
-            resolver.AddDirectories(new[] { new IncludeDirRule(@"=HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows Kits\Installed Roots\KitsRoot10;SubFolder") });
+            resolver.AddDirectories(new IncludeDirRule(@"=HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows Kits\Installed Roots\KitsRoot10;SubFolder"));
 
             using (LoggerMessageCountEnvironment(1, LogLevel.Error))
             using (LoggerMessageCountEnvironment(0, ~LogLevel.Error))
